Response to review
1. Removed used_station_count comment and added a doc-string note
   stating that it had been moved;
2. Refactored some utility functions into a utils submodule;
3. Changed accepted line-tags to a tuple to avoid mutation.
