+++ obspy/io/nordic/__init__.py
@@ -12,3 +12,11 @@ obspy.io.nordic - Nordic file format support for ObsPy
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 from future.builtins import *  # NOQA
+
+
+class NordicParsingError(Exception):
+    """
+    Internal general error for IO operations in obspy.core.io.nordic.
+    """
+    def __init__(self, value):
+        self.value = value
